swalert-options-delete

+++ FRONT/src/pages/Tareas.jsx
@@ -3,6 +3,7 @@ import { Container, Button, Box } from "@material-ui/core";
 import TareasCards from "../components/TareasCards";
 import { Link } from "react-router-dom";
 import axios from "axios";
+import Swal from "sweetalert2";
 
 const Tareas = () => {
 	const [tareas, setTareas] = useState({
@@ -29,9 +30,30 @@ const Tareas = () => {
 		setActualizar(false);
 	}, [actualizar]);
 
-	const borrar = async (id) => {
-		await axios.delete(`/api/v1/tareas/${id}`);
-		setActualizar(true);
+	const borrar = (id) => {
+
+		Swal.fire({
+			title: 'Esta seguro?',
+			text: "No podra recuperar!",
+			icon: 'warning',
+			showCancelButton: true,
+			confirmButtonColor: '#3085d6',
+			cancelButtonColor: '#d33',
+			confirmButtonText: 'Si, borrar!'
+		}).then(async (result) => {
+			if (result.value) {
+				await axios.delete(`/api/v1/tareas/${id}`);
+				Swal.fire({
+					icon: "success",
+					text: "Tarea eliminada...",
+					width: 250,
+					showConfirmButton: false,
+					timer: 2000,
+				})
+				setActualizar(true);
+			}
+		})
+	
 	};
 
 	const mostrar = () => {
